Problems 8-15
Did them all except 13 (which we were told to skip).

11_dictionary I wasn’t sure where to start with the printable function,
will look at answer.

12_rpn_calculator fails the RSpec test for raising the “calculator is
empty” error, yet outputs exactly that in Pry so the code works, it’s
apparently just not exactly what the test was looking for.

15_in_words passes the first test, but then fails when I added an if to
handle the second one. This makes no sense, as the if is definitely not
being triggered and should not affect the first test in any way. Will
look at answer to figure out what is going on.